I should make smaller commits, #255 and other stuff
Make (most?) flickering sprites flicker a bit slower #258
Attempt to mitigate #256, it's fixed for the booster on ceilings at least

+++ inc/stdint.h
@@ -4,11 +4,11 @@
 
 typedef signed char		int8_t;
 typedef signed short	int16_t;
-typedef signed int		int32_t;
+typedef signed long		int32_t;
 
 typedef unsigned char	uint8_t;
 typedef unsigned short	uint16_t;
-typedef unsigned int	uint32_t;
+typedef unsigned long	uint32_t;
 
 // SGDK Compatibility
